Mobile unit command support
