+++ core/assets/bundles/bundle.properties
@@ -305,6 +305,7 @@ ok = OK
 open = Open
 customize = Customize Rules
 cancel = Cancel
+command = Command
 openlink = Open Link
 copylink = Copy Link
 back = Back
