+++ core/src/mindustry/input/DesktopInput.java
@@ -12,7 +12,6 @@ import arc.scene.*;
 import arc.scene.ui.layout.*;
 import arc.util.*;
 import mindustry.*;
-import mindustry.ai.types.*;
 import mindustry.core.*;
 import mindustry.entities.units.*;
 import mindustry.game.EventType.*;
@@ -114,66 +113,9 @@ public class DesktopInput extends InputHandler{
             drawSelection(schemX, schemY, cursorX, cursorY, Vars.maxSchematicSize);
         }
 
-        if(commandMode){
-            //happens sometimes
-            selectedUnits.removeAll(u -> !u.isCommandable());
-
-            //draw command overlay UI
-            for(Unit unit : selectedUnits){
-                CommandAI ai = unit.command();
-                //draw target line
-                if(ai.targetPos != null){
-                    Position lineDest = ai.attackTarget != null ? ai.attackTarget : ai.targetPos;
-                    Drawf.limitLine(unit, lineDest, unit.hitSize / 2f, 3.5f);
-
-                    if(ai.attackTarget == null){
-                        Drawf.square(lineDest.getX(), lineDest.getY(), 3.5f);
-                    }
-                }
-
-                Drawf.square(unit.x, unit.y, unit.hitSize / 1.4f + 1f);
-
-                if(ai.attackTarget != null){
-                    Drawf.target(ai.attackTarget.getX(), ai.attackTarget.getY(), 6f, Pal.remove);
-                }
-            }
-
-            if(commandBuild != null){
-                Drawf.square(commandBuild.x, commandBuild.y, commandBuild.hitSize() / 1.4f + 1f);
-                var cpos = commandBuild.getCommandPosition();
-
-                if(cpos != null){
-                    Drawf.limitLine(commandBuild, cpos, commandBuild.hitSize() / 2f, 3.5f);
-                    Drawf.square(cpos.x, cpos.y, 3.5f);
-                }
-            }
-
-            if(commandMode && !commandRect){
-                Unit sel = selectedCommandUnit(input.mouseWorldX(), input.mouseWorldY());
-
-                if(sel != null && !(!multiSelect() && selectedUnits.size == 1 && selectedUnits.contains(sel))){
-                    drawCommand(sel);
-                }
-            }
-
-            if(commandRect){
-                float x2 = input.mouseWorldX(), y2 = input.mouseWorldY();
-                var units = selectedCommandUnits(commandRectX, commandRectY, x2 - commandRectX, y2 - commandRectY);
-                for(var unit : units){
-                    drawCommand(unit);
-                }
-
-                Draw.color(Pal.accent, 0.3f);
-                Fill.crect(commandRectX, commandRectY, x2 - commandRectX, y2 - commandRectY);
-            }
-        }
-
+        drawCommanded();
 
         Draw.reset();
-    }
-
-    public void drawCommand(Unit sel){
-        Drawf.square(sel.x, sel.y, sel.hitSize / 1.4f + Mathf.absin(4f, 1f), selectedUnits.contains(sel) ? Pal.remove : Pal.accent);
     }
 
     @Override
@@ -575,19 +517,7 @@ public class DesktopInput extends InputHandler{
 
         //select some units
         if(Core.input.keyRelease(Binding.select) && commandRect){
-            if(!tappedOne){
-                var units = selectedCommandUnits(commandRectX, commandRectY, input.mouseWorldX() - commandRectX, input.mouseWorldY() - commandRectY);
-                if(multiSelect()){
-                    //tiny brain method of unique addition
-                    selectedUnits.removeAll(units);
-                }else{
-                    //nothing selected, clear units
-                    selectedUnits.clear();
-                }
-                selectedUnits.addAll(units);
-                commandBuild = null;
-            }
-            commandRect = false;
+            selectUnitsRect();
         }
 
         if(Core.input.keyTap(Binding.select) && !Core.scene.hasMouse()){
@@ -711,11 +641,6 @@ public class DesktopInput extends InputHandler{
         }
     }
 
-    //TODO when shift is held? ctrl?
-    public boolean multiSelect(){
-        return false;
-    }
-
     @Override
     public boolean tap(float x, float y, int count, KeyCode button){
         if(scene.hasMouse() || !commandMode) return false;
@@ -754,30 +679,7 @@ public class DesktopInput extends InputHandler{
         if(scene.hasMouse() || !commandMode) return false;
 
         if(button == KeyCode.mouseRight){
-            //right click: move to position
-
-            //move to location - TODO right click instead?
-            Vec2 target = input.mouseWorld().cpy();
-
-            if(selectedUnits.size > 0){
-
-                Teamc attack = world.buildWorld(target.x, target.y);
-
-                if(attack == null || attack.team() == player.team()){
-                    attack = selectedEnemyUnit(target.x, target.y);
-                }
-
-                int[] ids = new int[selectedUnits.size];
-                for(int i = 0; i < ids.length; i++){
-                    ids[i] = selectedUnits.get(i).id;
-                }
-
-                Call.commandUnits(player, ids, attack instanceof Building b ? b : null, attack instanceof Unit u ? u : null, target);
-            }
-
-            if(commandBuild != null){
-                Call.commandBuilding(player, commandBuild, target);
-            }
+            commandTap(x, y);
         }
 
         return super.touchDown(x, y, pointer, button);
